requirements.txt
# all requirements needed by the project
pytest-playwright
pytest-json-report